Moved AccountConstants up a level.
