+++ src/main/kotlin/org/wfanet/measurement/api/BUILD.bazel
@@ -6,6 +6,16 @@ package(default_visibility = [
 ])
 
 kt_jvm_library(
+    name = "account_constants",
+    srcs = ["AccountConstants.kt"],
+    deps = [
+        "//src/main/proto/wfa/measurement/internal/kingdom:account_kt_jvm_proto",
+        "@wfa_common_jvm//imports/java/io/grpc:api",
+        "@wfa_common_jvm//imports/java/io/grpc:context",
+    ],
+)
+
+kt_jvm_library(
     name = "public_api_version",
     srcs = ["Version.kt"],
     deps = [],
