+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/BUILD.bazel
@@ -25,13 +25,3 @@ kt_jvm_library(
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
     ],
 )
-
-kt_jvm_library(
-    name = "account_constants",
-    srcs = ["AccountConstants.kt"],
-    deps = [
-        "//src/main/proto/wfa/measurement/internal/kingdom:account_kt_jvm_proto",
-        "@wfa_common_jvm//imports/java/io/grpc:api",
-        "@wfa_common_jvm//imports/java/io/grpc:context",
-    ],
-)
