+++ src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha/BUILD.bazel
@@ -13,7 +13,7 @@ kt_jvm_library(
     name = "account_authentication_server_interceptor",
     srcs = ["AccountAuthenticationServerInterceptor.kt"],
     deps = [
-        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:account_constants",
+        "//src/main/kotlin/org/wfanet/measurement/api:account_constants",
         "//src/main/kotlin/org/wfanet/measurement/common/identity",
         "//src/main/proto/wfa/measurement/internal/kingdom:accounts_service_kt_jvm_grpc",
         "//src/main/proto/wfa/measurement/internal/kingdom:accounts_service_kt_jvm_proto",
